fix: Componentiza estilos.
